Updated train_erac.py

+++ mt/erac/train_erac.py
@@ -201,7 +201,7 @@ def train_mle(src, tgt):
     mask = tgt[1:].ne(tgt_pad_idx).float()
 
     log_act_dist = actor(src, tgt)
-    nll = -log_act_dist.gather(2, tgt[1:].unsqueeze(2)).squeeze(2) * mask
+    nll = -log_act_dist.gather(2, tgt[1:].unsqueeze(2).to(torch.int64)).squeeze(2) * mask
 
     loss = nll.sum(0).mean()
